Add env var to schedule cleanupInactiveUsers job and refactor
